requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
apify-client>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
